rmtree fix

+++ pydmc/SubmitTools.py
@@ -194,7 +194,8 @@ class SubmitTools(object):
                 # (1) create calc_dir
                 tag = '%s-%s' % (xc, calc)
                 calc_dir = os.path.join(self.launch_dir, tag)
-                if fresh_restart:
+
+                if os.path.exists(calc_dir) and fresh_restart:
                     rmtree(calc_dir)
                 if not os.path.exists(calc_dir):
                     os.mkdir(calc_dir)
